browser/components/resistfingerprinting/test/mochitest/test_animation_api.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1382545
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1382545</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1382545 **/
  SimpleTest.waitForExplicitFinish();

  var currentPref = "";
  var expectedPrecision = 0;
  window.onload = () => {
    currentPref = "privacy.resistFingerprinting";
    expectedPrecision = 100000;
    SpecialPowers.pushPrefEnv({"set":
      [
        ["privacy.resistFingerprinting", true],
        ["dom.animations-api.core.enabled", true],
        ["privacy.reduceTimerPrecision", false],
        ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
      ]
    }, runTest);
  };

  function runTest() {
    window.open("file_animation_api.html");
  }

  let completed = 0;
  const numTests = 8;
  function done() {
    completed++;
    if (completed == numTests) {
      SimpleTest.finish();
    } else {
      nextTest();
    }
  }

  function nextTest() {
    // ----------------------------------------------
    if (completed == 1) {
      currentPref = "privacy.reduceTimerPrecision";
      expectedPrecision = 100000;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", false],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", true],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else if (completed == 2) {
      currentPref = "privacy.resistFingerprinting";
      expectedPrecision = 50000;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", true],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", false],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else if (completed == 3) {
      currentPref = "privacy.reduceTimerPrecision";
      expectedPrecision = 50000;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", false],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", true],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else if (completed == 4) {
      currentPref = "privacy.resistFingerprinting";
      expectedPrecision = 100;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", true],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", false],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else if (completed == 5) {
      currentPref = "privacy.reduceTimerPrecision";
      expectedPrecision = 100;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", false],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", true],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else if (completed == 6) {
      currentPref = "privacy.resistFingerprinting";
      expectedPrecision = 13;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", true],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", false],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else if (completed == 7) {
      currentPref = "privacy.reduceTimerPrecision";
      expectedPrecision = 13;
      SpecialPowers.pushPrefEnv({"set":
        [
          ["privacy.resistFingerprinting", false],
          ["dom.animations-api.core.enabled", true],
          ["privacy.reduceTimerPrecision", true],
          ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision]
        ]
      }, runTest);
    // ----------------------------------------------
    } else {
      ok(false, "I seem to have asked for " + numTests +
         " tests, but don't know how to run them all.");
      SimpleTest.finish();
    }
  }
  </script>
</head>
<body>
</body>
</html>
